requirements.txt
numpy == 1.15.2
pillow == 5.2.0
scikit-image == 0.14.2
scipy == 1.2.1
torch == 1.1.0
torchvision == 0.3.0
psutil == 5.4.7
pandas == 0.24.0
bitstring == 3.1.5
matplotlib == 3.0.2
setproctitle == 1.1.10
